Added more NitroTPM details on startup (#101)

+++ crates/vouch-server/src/db/document_type.rs
@@ -83,7 +83,7 @@ pub struct Document<T> {
     /// Optional expiration timestamp.
     pub expires_at: Option<Timestamp>,
     /// Optimistic concurrency version. Incremented on every update.
-    pub version: i64,
+    pub version: i32,
     /// Lightweight last-used timestamp (column-level, no encrypt/decrypt).
     pub last_used_at: Option<Timestamp>,
 }
